# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ML API",
    "description": "\n    Machine Learning API\n\n    - Upload datasets\n    - Register and train ML models\n    - Evaluate and rank models\n    - Make predictions\n    - List models and view reports\n    ",
    "contact": {
      "name": "AM",
      "url": "https://github.com/a1mart/ml",
      "email": "aidan3martin@gmail.com"
    },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "version": "1.0.0"
  },
  "paths": {
...
    "/models/register/{dataset_name}": {
      "post": {
        "tags": ["Models"],
        "summary": "Register Model",
        "description": "Register a model for a specific dataset",
        "operationId": "register_model_models_register__dataset_name__post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModelRegister" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/models/train/{dataset_name}": {
      "post": {
        "tags": ["Models"],
        "summary": "Train Models",
        "description": "FIXED: Start training models asynchronously with unified request structure",
        "operationId": "train_models_models_train__dataset_name__post",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TrainingRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TrainingResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/models/tasks/{task_id}": {
      "get": {
        "tags": ["Models"],
        "summary": "Get Task Status",
        "description": "Get the status of a training task",
        "operationId": "get_task_status_models_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Task Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskStatus" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/models/list/{dataset_name}": {
      "get": {
        "tags": ["Models"],
        "summary": "List Models",
        "description": "List all models (registered and trained) for a dataset",
        "operationId": "list_models_models_list__dataset_name__get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/models/trained/{dataset_name}/{model_id}": {
      "delete": {
        "tags": ["Models"],
        "summary": "Delete Trained Model",
        "description": "Delete a trained model",
        "operationId": "delete_trained_model_models_trained__dataset_name___model_id__delete",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          },
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Model Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelRegister": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "model_name": { "type": "string", "title": "Model Name" },
          "params": {
            "anyOf": [{ "type": "object" }, { "type": "null" }],
            "title": "Params",
            "default": {}
          },
          "target_column": { "type": "string", "title": "Target Column" },
          "drop_columns": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Drop Columns",
            "default": []
          }
        },
        "type": "object",
        "required": ["name", "model_name", "target_column"],
        "title": "ModelRegister"
      },
      "TaskStatus": {
        "properties": {
          "task_id": { "type": "string", "title": "Task Id" },
          "status": { "type": "string", "title": "Status" },
          "progress": { "type": "number", "title": "Progress" },
          "start_time": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "End Time"
          },
          "error_message": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Error Message"
          },
          "results": {
            "anyOf": [{ "type": "object" }, { "type": "null" }],
            "title": "Results"
          }
        },
        "type": "object",
        "required": ["task_id", "status", "progress"],
        "title": "TaskStatus"
      },
      "TrainingRequest": {
        "properties": {
          "target_column": { "type": "string", "title": "Target Column" },
          "drop_columns": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Drop Columns",
            "default": []
          },
          "n_splits": { "type": "integer", "title": "N Splits", "default": 5 },
          "random_state": {
            "type": "integer",
            "title": "Random State",
            "default": 1
          },
          "save_format": {
            "type": "string",
            "title": "Save Format",
            "default": "pkl"
          },
          "model_ids": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Model Ids"
          }
        },
        "type": "object",
        "required": ["target_column"],
        "title": "TrainingRequest",
        "description": "FIXED: Combined training request model"
      },
      "TrainingResponse": {
        "properties": {
          "task_id": { "type": "string", "title": "Task Id" },
          "status": { "type": "string", "title": "Status" },
          "message": { "type": "string", "title": "Message" },
          "poll_url": { "type": "string", "title": "Poll Url" }
        },
        "type": "object",
        "required": ["task_id", "status", "message", "poll_url"],
        "title": "TrainingResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}